+++ libraries/base/Foreign/Ptr.hs
@@ -7,7 +7,7 @@
 -- Module      :  Foreign.Ptr
 -- Copyright   :  (c) The FFI task force 2001
 -- License     :  BSD-style (see the file libraries/base/LICENSE)
--- 
+--
 -- Maintainer  :  ffi@haskell.org
 -- Stability   :  provisional
 -- Portability :  portable
@@ -41,12 +41,15 @@ module Foreign.Ptr (
     -- Free the function pointer created by foreign export dynamic.
 
     -- * Integral types with lossless conversion to and from pointers
-    IntPtr,
+    IntPtr(..),
     ptrToIntPtr,
     intPtrToPtr,
-    WordPtr,
+    WordPtr(..),
     ptrToWordPtr,
     wordPtrToPtr
+
+    -- See Note [Exporting constructors of marshallable foreign types]
+    -- for why the constructors for IntPtr and WordPtr are exported.
  ) where
 
 import GHC.Ptr
@@ -97,3 +100,21 @@ ptrToIntPtr (Ptr a#) = IntPtr (I# (addr2Int# a#))
 -- | casts an @IntPtr@ to a @Ptr@
 intPtrToPtr :: IntPtr -> Ptr a
 intPtrToPtr (IntPtr (I# i#)) = Ptr (int2Addr# i#)
+
+{-
+Note [Exporting constructors of marshallable foreign types]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+One might expect that IntPtr, WordPtr, and the other newtypes in the
+Foreign.C.Types and System.Posix.Types modules to be abstract, but this is not
+the case in GHC (see Trac #5229 and #11983). In fact, we deliberately export
+the constructors for these datatypes in order to satisfy a requirement of the
+Haskell 2010 Report (§ 8.4.2) that if a newtype is used in a foreign
+declaration, then its constructor must be visible.
+
+This requirement was motivated by the fact that using a type in a foreign
+declaration necessarily exposes some information about the type to the user,
+so being able to use abstract types in a foreign declaration breaks their
+abstraction (see Trac #3008). As a result, the constructors of all FFI-related
+newtypes in base must be exported in order to be useful for FFI programming,
+even at the cost of exposing their underlying, architecture-dependent types.
+-}