+++ testsuite/tests/ffi/should_compile/T11983.c
@@ -1,0 +1,5 @@
+#include <stdint.h>
+#include "T11983.h"
+
+void intptr_example(intptr_t i)   {}
+void uintptr_example(uintptr_t u) {}