+++ testsuite/tests/ffi/should_compile/T11983.h
@@ -1,0 +1,9 @@
+#ifndef T11983_H
+#define T11983_H
+
+#include <stdint.h>
+
+void intptr_example(intptr_t);
+void uintptr_example(uintptr_t);
+
+#endif // T11983_H